+++ src/tools/delegate-task/index.ts
@@ -1,3 +1,4 @@
-export { createDelegateTask, type DelegateTaskToolOptions } from "./tools"
+export { createDelegateTask, resolveCategoryConfig, buildSystemContent } from "./tools"
+export type { DelegateTaskToolOptions, SyncSessionCreatedEvent, BuildSystemContentInput } from "./tools"
 export type * from "./types"
 export * from "./constants"
